+++ src/main/java/com/example/assignment_2/bussiness/controller/LaboratoryController.java
@@ -8,6 +8,7 @@ import com.example.assignment_2.bussiness.model.base.Laboratory;
 import com.example.assignment_2.bussiness.model.base.Teacher;
 import com.example.assignment_2.bussiness.model.create.LaboratoryCreateModel;
 import com.example.assignment_2.bussiness.service.implementation.LaboratoryServiceImplementation;
+import com.example.assignment_2.bussiness.service.implementation.TeacherServiceImplementation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -22,6 +23,9 @@ public class LaboratoryController {
 
     @Autowired
     private LaboratoryServiceImplementation laboratoryService;
+    @Autowired
+    private TeacherServiceImplementation teacherService;
+
 
     @GetMapping("/getAll")
     public ResponseEntity<List<LaboratoryDTO>> findAll() {
@@ -45,14 +49,25 @@ public class LaboratoryController {
     }
 
     @PostMapping("/create")
-    public ResponseEntity<LaboratoryDTO> create(@RequestBody LaboratoryCreateModel createModel) {
+    public ResponseEntity<LaboratoryDTO> create(@RequestBody LaboratoryCreateModel createModel, @RequestHeader("Token") String authenticationToken) {
+        Teacher teacher = teacherService.findByAuthenticationToken(authenticationToken);
+
+        if(teacher == null){
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        }
+
         Laboratory saved = laboratoryService.create(createModel);
 
         return new ResponseEntity<>(new LaboratoryDTO(saved), HttpStatus.OK);
     }
 
     @DeleteMapping("/deleteById/{id}")
-    public ResponseEntity<LaboratoryDTO> deleteById(@PathVariable Long id) {
+    public ResponseEntity<LaboratoryDTO> deleteById(@PathVariable Long id, @RequestHeader("Token") String authenticationToken) {
+        Teacher teacher = teacherService.findByAuthenticationToken(authenticationToken);
+
+        if(teacher == null){
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        }
         Laboratory deleted = laboratoryService.deleteById(id);
 
         if(deleted == null){
@@ -63,7 +78,13 @@ public class LaboratoryController {
     }
 
     @PostMapping("/updateById/{id}")
-    public ResponseEntity<LaboratoryDTO> update(@PathVariable Long id, @RequestBody LaboratoryCreateModel newValue) {
+    public ResponseEntity<LaboratoryDTO> update(@PathVariable Long id, @RequestBody LaboratoryCreateModel newValue, @RequestHeader("Token") String authenticationToken) {
+        Teacher teacher = teacherService.findByAuthenticationToken(authenticationToken);
+
+        if(teacher == null){
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        }
+
         Laboratory updated = laboratoryService.update(id, newValue);
 
         if(updated == null){
@@ -73,7 +94,14 @@ public class LaboratoryController {
         return new ResponseEntity<>(new LaboratoryDTO(updated), HttpStatus.OK);
     }
 
-    public ResponseEntity<List<AttendanceDTO>> getAttendance(Long id) {
+    @GetMapping("/getAttendance/{id}")
+    public ResponseEntity<List<AttendanceDTO>> getAttendance(@PathVariable Long id, @RequestHeader("Token") String authenticationToken) {
+        Teacher teacher = teacherService.findByAuthenticationToken(authenticationToken);
+
+        if(teacher == null){
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        }
+
         List<Attendance> attendances = laboratoryService.getAttendance(id);
 
         if(attendances == null){
